Revert small change

+++ esphome/components/tuya/climate/tuya_climate.cpp
@@ -119,7 +119,7 @@ void TuyaClimate::control(const climate::ClimateCall &call) {
 
     ESP_LOGV(TAG, "Setting target temperature: %.1f", target_temperature);
     this->parent_->set_integer_datapoint_value(*this->target_temperature_id_,
-                                               target_temperature / this->target_temperature_multiplier_);
+                                               (int) (target_temperature / this->target_temperature_multiplier_));
   }
 
   if (call.get_preset().has_value()) {
